webserv

+++ inc/ServerBlock.hpp
@@ -6,7 +6,7 @@
 /*   By: nelidris <nelidris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/13 18:44:37 by nelidris          #+#    #+#             */
-/*   Updated: 2023/05/13 18:44:53 by nelidris         ###   ########.fr       */
+/*   Updated: 2023/06/08 14:00:54 by nelidris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@
 
 struct ServerBlock
 {
-	std::pair<size_t, short> listen; // [ IP_ADDR, PORT ]
+	std::pair<int, short> listen; // [ IP_ADDR, PORT ]
 	std::vector<std::string> server_names;
-	size_t client_max_body_size; // IN BYTES ONLY
+	ssize_t client_max_body_size; // IN BYTES ONLY
 	std::map<short, std::string> error_pages; // [ PATH, STATUS_CODE ]
 	std::map<std::string, LocationBlock> locations; // [ URI, LOCATION_BLOCK ]
